renamed `controllers` to `controller`
