+++ joey/tests/test_resources.py
@@ -14,7 +14,7 @@ from joey import server
 from joey import util
 from joey.server import db
 from joey.constants import APPNAME, PACAKAGE_FOLDER, VERSION
-from joey.server.controllers.fixture import load
+from joey.server.controller.fixture import load
 from joey.server.models.base import Database
 
 log = logging.getLogger(__name__.split('.')[-1])
