+++ joey/util/cli/develop.py
@@ -13,7 +13,7 @@ from traitlets.config import get_config
 
 from joey import server, util, constants
 from joey.server import db, shell
-from joey.server.controllers import fixture
+from joey.server.controller import fixture
 from joey.util.context import ( get_config_location, 
     select_configuration_questionaire, configuration_wizard )
